Moved tstep_counter increment
The previous commit incremented model%numerics%tstep_counter in the standalone driver,
cism_run_dycore.  This will not work for coupled CESM runs, so I moved the increment
to the beginning of subroutines glide_tstep_p1 and glissade_tstep.

The counter is initially set to 0 in the respective initialize routines.
In a subsequent commit, Bill Sacks will add tstep_counter to the restart file,
so that it can be read at restart. After that commit, exact restart should work
for isostasy in both standalone and coupled runs.