Increment tstep_count for evolve_ice false
With ice evolution turned on, tstep_count gets incremented inside the
dycore's tstep subroutine. Since we don't call the tstep subroutine with
evolve_ice false, we need to increment it in glad_i_tstep_gcm.

In practice, this shouldn't matter, but it's needed so that CESM's exact
restart tests pass, since tstep_count is written to CISM's history
files.

This makes me wonder if we should go back to just having tstep_count
incremented outside the dycore, as was done before Bill Lipscomb's
commit 1d16f81a ("Moved tstep_counter increment"). But for now, I'm
going to leave this as is - with tstep_count incremented in the dycore
in the evolving ice case, but outside the dycore in the non-evolving
case.