Initialize usrf consistently with what's done in the run loop
In the run loop, updates to usrf guarantee that it ends up at least
0. This commit changes the initialization of usrf to have that guarantee
as well.

I originally tried this hoping it would fix the isostasy exact restart
problems. It didn't (a previous commit fixed those problems instead),
but Bill Lipscomb and I felt it was best to be consistent in this way
anyway - even though, in practice, these changes might never make a
difference.