Moved the isostasy call to the start of glissade_tstep
In Glide, isostasy is computed at the end of the time step, in glide_tstep_p3.
When glissade_tstep was written, isostasy was left near the end of the time step,
after transport and calving but before the diagnostic velocity solve.

It turns out to be problematic to compute isostasy between calving and velocity.
Consider, for instance, the marine_margin option that removes all floating ice.
Suppose we have a marine-based cell that is just barely grounded after transport.
Then it will be left in place by the calving.  But if the topography is sinking,
the cell can switch from grounded to floating during the isostasy calculation.
Then the velocity calculation takes the cell to be floating and potentially
can compute large velocities as a result of having beta = 0 at a neighboring vertex.
This happened in a recent simulation by Sarah Bradley.

In order to maintain a self-consistent ice state conducive to a stable velocity solve,
it is better to do the isostatic adjustment after the velocity solve. Since the velocity
solve is done at the end of the time step, this means that isostasy should be computed
at the start of the next time step. I think this also resolves Matt Hoffman's concern
that the isostasy call was in the wrong place for consistent forward differencing.

With this commit, subroutine glissade_isostasy_solve is called at the start of glissade_tstep,
just after updating the internal clock.  The Glissade physics sequence is now as follows:

At initialization:
- Diagnose velocity

During each time step:
- Isostasy
- Thermal solve
- Transport
- Calving
- Second thermal solve (optional)
- Diagnose velocity

I left Glide unchanged, since Glide does not support marine sheet simulations where these issues arise.

I verified that restart remains exact for a standalone 4-km Greenland test.
Bill S. can verify exact restart for CESM runs.

Answers will change slightly for Glissade runs with isostasy, but are BFB for runs without isostasy.