Changed the logic for limiting surface gradients to max_slope
To help with stability for simulations with steep topography, the x and y components
of the surface elevation gradient can be limited to a user-prescribed value, max_slope.
We have been running Greenland simulations with max_slope = 0.10.
However, the logic was flawed. This commit fixes the logic.

Here is the old logic:
(1) Compute gradients df/dx and df/dy on edges.
(2) Limit the magnitude of edge gradient components not to exceed max_slope.
(3) Average the edge gradient components to vertices.

Here is the new logic:
(1) Compute gradients df/dx and df/dy on edges.
(2) Average the edge gradient components to vertices.
(3) Limit the magnitude of vertex gradient components not to exceed max_slope.

The problem with the old logic is that if gradients on each side of a vertex
are large and of opposite sign, they can average to zero at the vertex.
This led to an unphysical buildup of ice in one cell in a recent Greenland run.
The change in logic prevents the buildup.

The fix was made in subroutines glissade_centered_gradient and glissade_upstream_gradient
in glissade_grid_operators.F90.

Answers will change for any simulation where gradients are limited to max_slope.
This includes standard Greenland simulations.