Added overwrite_acab options
For standalone runs it can be helpful to overwrite the SMB (usually with a negative value)
where (1) the input SMB is zero (e.g., outside the ice sheet mask where RACMO computes SMB)
or (2) the initial ice thickness is below some threshold (to inhibit ice advance
into cells that initially have zero or thin ice).

Option (1) has previously been supported with somewhat different logic.
Option (2) is added here, as a way of inhibiting ice advance into grid cells
where RACMO provides a positive SMB but we do not want the ice sheet to advance
(because these cells are covered by peripheral glaciers rather than the ice sheet).

The new option is called overwrite_acab. The options are:
    OVERWRITE_ACAB_NONE = 0 (the default)
    OVERWRITE_ACAB_ZERO_ACAB = 1
    OVERWRITE_ACAB_THCKMIN = 2

The user-specified acab, called overwrite_acab_value, is a scalar with a fixed value.
It is written to cells with overwrite_acab_mask = 1.
The mask is computed at initialization depending on the value of overwrite_acab.
It is written to restart files so that it does not need to be recomputed.

For option 1, any cell with an input value acab = 0 is assigned overwrite_acab_mask = 1.

For option 2, any cell with initial thickness <= overwrite_acab_minthck has overwrite_acab_mask = 1.
The default value is overwrite_acab_minthck = 0.0, but the user can set this value in the config file.
For standalone Greenland spinups I am setting it to 5 m, because many cells have an input thickness of ~1 m.
I am assuming that these cells are not part of the initial ice sheet.

For coupled runs it is probably advisable to stick to overwrite_acab = 0, since CESM
provides an SMB throughout the CISM domain. (But I can imagine a scenario where
a user might want to overwrite the CESM value in certain cells to prevent ice advance.)

I checked that the model restarts exactly with overwrite_acab = 1 and 2.

Answers will not change for the default option, overwrite_acab = 0.