Added a diagnostic basal temperature array
It is often useful to plot the basal temperature, btemp.
This field can be listed as 'btemp' in the config file and written to output.
Until now, it has been a copy of the upn level of the 3D temp array.

However, ice-free columns currently are assigned a temperature of min(artm,0)
at all levels, including the basal surface. When plotting btemp in output,
there is thus a mix between physical temperatures where there is ice, and
artm values where there is no ice.  This is not ideal for visualization.

One solution would be to assign T = 0 as the vertical profile of ice-free columns.
However, this would be answer-changing for the dome and similar problems.
(I'm not sure exactly why; I suspect that when new columns come into existence,
their temperature profile is borrowed from the previous ice-free profile instead
of being created from scratch.)

To avoid answer changes at this stage, I simply created a new diagnostic array
called btemp in the temper derived type. It is a copy of temp(upn,:,:),
except that btemp = 0 where H = 0.
It is automatically diagnosed from temp(upn) and thck for both glide and glissade runs.

At some point I would like to make the surface and basal temperatures 2D prognostic
fields, with the 3D temp array applying only to internal layers. This is the
MPAS landice convention.

Since the new field is diagnostic only, this commit is BFB.