Set N = 0 for floating cells in effective pressure calculations
Currently there are five options for computing the effective pressure N,
which is a key term in the Coulomb and pseudoplastic sliding laws.
Three of these--HO_EFFECPRESS_BPMP, HO_EFFECPRESS_BMLT and HO_EFFECPRESS_BWAT--
are similar in spirit; they are based on the idea that N decreases when
the bed thaws and water is present.

With this commit, N is set to zero for floating cells for these three
effective pressure options. This makes sense physically, since floating ice
is entirely supported by water at the bed (at least in hydrostatic flow
approximations).

However, N is not automatically set to zero in floating cells for
HO_EFFECPRESS_OCEAN_PENETRATION, because this would be problematic when ocean_p = 0.
(See code comments for an explanation.)

Also, we retain N = rhoi*g*H everywhere (including floating cells) for
HO_EFFECPRESS_OVERBURDEN. This is not physically realistic, but the point
of this option is to see what happens when N = overburden everywhere.

Answers are changed for the BPMP, BMLT and BWAT options in any simulations with floating ice.
Answers are unchanged in simulations without floating ice (e.g., where all floating ice
immediately calves).

With this change, more linear iterations are now required for convergence in simulations
with floating ice and a pseudoplastic sliding law. Presumably this is related to the
lower values of N near the grounding line when including N = 0 values in spatial averages.

The slower convergence described in the last two commits suggests that it might not
be desirable to run Greenland simulations with floating ice, unless we
are particularly interested in resolving small ice shelves and ice tongues.

For Antarctic simulations with large areas of floating ice, convergence
of the linear solver will likely be slow until we have better preconditioning.