Added eigencalving option; worked on plume model
This commit brings in some work that was done on the misomip branch a couple of months ago
but not checked in:

(1) a new eigencalving option based on the eigencalving scheme of Albrecht and Levermann.
This scheme differs from A & L in that instead of computing a lateral calving retreat rate,
it computes a 2D threshold thickness field to decide where to calve.  The threshold thickness
in a given cell is based on the product of principal stresses of the horizontal strain rate tensor.

I added some I/O fields associated with the strain rate tensor to help with testing.

This scheme has been implemented but needs more testing, especially for Antarctica.

(2) substantial changes in glissade_bmlt_float.F90, mainly for the plume model.
I am still trying to solve for plume thickness in an efficient, robust way without
oscillations in the free surface height. This work remains in progress.

The standard LIVV tests pass as expected. However, some more testing will be needed
to make sure other recently added capabilities have not broken.