Modifications to the new hybrid gradient option
Here I have repeated some modifications associated with commit 194dc4c
in the latest greenland_cleanup branch. I have not tried to rebase to that branch,
because then some of the desired changes could be undone by commit 5060379 (which took
place earlier in time, but would be later in the history). I am hoping that
once this commit is in place, a future rebase will end up at the desired end point.