Changes to read bmlt_float from an external file; mask bug fix
For initMIP-Antarctica spin-up experiments, a prescribed basal melt rate for floating ice
is read from the input file. For sensitivity experiments, an additional bmlt anomaly
field is read from the input file.

With this commit, there is a new option, whichbmlt_float = BMLT_FLOAT_EXTERNAL = 4,
which is set in the config file when reading in an external bmlt_float field.
(This option could also be used for bmlt_float provided by the coupler.)

Another new option, enable_bmlt_float_anomaly, is false by default, but can be set to true
in the config file to read in a prescribed anomaly field.

Also, I fixed two bugs in mask calls in glissade.F90.  Subroutine glissade_get_masks
has three optional masks: floating_mask, ocean_mask and land_mask.  If not all masks are called,
the ones that are called need to be identified explicitly in the argument list
(e.g., 'land_mask = land_mask'). In the previous code, I was inadvertently passing floating_mask
instead of ocean_mask to the transport solver, which prevented accumulation in floating cells.

Detailed changes:

- For whichbmlt_float = 4, the bmlt_float_external field is loaded (if present in the input file)
and is added to the restart file.

- I added a bmlt_ground field so that the grounded and floating basal melt rates
can be cleanly separated in diagnostics.  As before, bmlt = bmlt_ground + bmlt_float.

- I put all the bmlt fields in a new derived type, glide_basal_melt. Previously, some were
in the temper type and others in the bmltfloat type. The bmltfloat type no longer exists.

- I moved plume-related fields from the bmltfloat type to a new glide_plume derived type.

- For consistency, I created a new option called enable_acab_anomaly, which works
the same way as enable_bmlt_float_anomaly.

- I removed the flux_correction field, which was redundant relative to acab_anomaly.

- In glissade_tstep I added a subroutine called glissade_bmlt_float_solve, in analogy
to glissade_transport_solve, glissade_calving_solve, etc.

- I replaced subroutine glissade_add_acab_anomaly with glissade_add_mbal_anomaly,
which is functionally equivalent but now is used for both acab and bmlt anomalies.

- I renamed parameter bmlt_float_rate as bmlt_float_const, and modified
mismip+.config.template accordingly.

- I made a minor correction to the mismip+ README file.

I verified that answers are unchanged in standard LIVV tests.
Results look sensible for the MISMIP+ Ice0, Ice1 and Ice2 experiments.