Added a thklim_float threshold for numerical stability; eigencalving changes
I have been running Antarctic experiments while testing thickness-based calving
and eigencalving.  When thin floating ice (thck ~ thklim ~1-10 m) is dynamically active,
velocities can be large near the calving margin, especially when there is a ring of thin
floating cells protected from calving (as required for calving fronts to advance
and retreat).  The problem is that there can be large surface elevation gradients
between thick floating ice (in the main part of the shelf) and thin floating ice
(in the protected ring). These gradients can drive large and oscillating velocities.

The solution implemented here is to introduce a parameter called thklim_float, which is
analogous to thklim but applies only to floating ice.  Floating cells thinner than thklim_float
are dynamically inactive and are treated as ice-free ocean cells for gradient computations.
Then the gradient between thick (active) floating ice and thin (inactive) floating ice is zero.

The glissade_get_masks subroutine now has an optional thklim_float parameter.
It should be called with this parameter whenever thklim_float /= thklim.
Without this parameter, thklim_float defaults to thklim.  For some mask calls, we want
ice_mask = 1 for any cell, grounded or floating, with thck > 0 (rather than thck > thklim).
For these calls we pass a value of 0.0 for thklim; thklim_float then defaults to 0.
I went through the various calls to glissade_get_masks and added thklim_float as needed.

For thickness-based calving, thklim_float should have the same value as calving_minthck.
This will inhibit large gradients near the calving front.  For flexibility, thklim_float
and calving_minthck remain separate parameters.  But when running thickness-based calving,
thklim_float is automatically set to calving_minthck at startup (with a message in the log file).

With thklim_float = calving_minthck = 100 m, I verified that max velocities are ~4 km/yr
or less for the first 20 years of an initMIP Antarctic spinup.

I also worked on the eigencalving scheme.  One problem was that there was horizontal
convergence instead of spreading in thin floating cells on the marine margin.  Setting
thklim_float = calving_minthck largely solved this problem.

I then simplified the eigencalving, making it more similar to Levermann et al. (2012).
Previously, the spreading rate (i.e., the determinant of the horizontal strain-rate tensor)
was converted to a minimum thickness.  Now, the spreading rate is converted to a
lateral calving rate (as in Levermann), which in turn is converted to a thinning rate
(unlike Levermann) with an equivalent volume loss given the grid cell size. That is,
the calving rate C is given by

    C = K * eps+ * eps-

where C has units of m/yr, K has units of m*yr, and eps+ and eps- are eigenvalues
of the strain rate tensor with units of yr^-1.  This follows Levermann.

Then in a given cell of thickness H, the thinning rate m is based on conservation
of calved volume:

     m * dx = C * H

where dx is the grid cell size.

The eigencalving then becomes a supplement to thickness-based calving.
There is a specified calving thickness threshold (e.g., calving_minthck = 100 m).
Continual spreading near the calving front causes cells to thin below the threshold,
balanced by inflow of thicker ice from upstream.  The balance of processes determines
the steady-state calving front location.

I tested the eigencalving with different values of K and calving_minthck.
The flow seems stable, without large velocities, for K = 0.5e8 to 1.0e9 m*yr.
A large spreading rate induces large thinning, which reduces the spreading rate and thinning.
So the thinning rates tend to stabilize at modest values.

Other changes:

- I changed calving_minthck from a 2D field back to a parameter.

- For eigencalving, I added the strain_rate_determinant field to the restart file.
On restart, the velocities alone are not sufficient to compute this determinant
(because it is derived from the strain-rate tensor, which is derived from the
stress tensor computed in the HO solver). I verified exact restart for eigencalving.

- I changed the name of the 'calving' output field to 'calving_thck', since it is
a thickness rather than a rate.

- I removed the parameter floating_path_minthck from remove_floating_islands.
Instead, thklim_float is passed in.  Since floating cells thinner than thklim_float
are dynamically inactive, paths through these cells are excluded as paths
back to grounded ice (which was the purpose of floating_path_minthck).

- I improved an instruction in README.mismip+ (thanks to Gunter).

This is an answer-changing commit for marine ice simulations, because of the
new thklim_float parameter.  Standard LIVV tests still pass.