Changes and simplifications in the plume model
When I worked on the plume model in the spring, I was trying to consistently solve
the velocity and continuity equations with a free-surface term in the momentum balance.
I was unable to get robust, stable solutions in regions of complex topography.

Following a conversation with Alistair Adcroft, I decided to take a different approach.
Instead of introducing a free-surface term, I went back to a steady-state balance between
the pressure-gradient force, Coriolis and basal drag, supplemented by either
(1) logic that effectively reduces the Coriolis force in cells adjacent to the boundary, or
(2) a Laplacian horizontal friction term that effectively enforces a no-penetration BC
    for flow into the boundary walls and a free-slip condition for flow parallel to walls.

I implemented (1), which works well in the sense that it allows consistent solutions of
melting, velocity and continuity, with a rightward velocity in the north row of cells.
However, this method will not converge with increasing resolution, because the logic is based
on whether a given cell is adjacent to a wall (not the distance from the wall).

I also implemented (2), which shows some promise but needs more work. Following a converged
velocity solve, the lateral drag terms seem to be in the right direction near the boundaries.
However, if I naively add then to the momentum balance, the solution is unstable.
Maybe the lateral drag term will work better if, for example, applied to an already-converged flow
as an explicit forcing term with a small timestep.  To be continued.

In either case, the free-surface term is zero, which means the plume no longer knows whether
it is thick enough to reach the sea floor. The ocean layer below the plume is treated
as infinitely thick everywhere.

Other changes:

- Removed some code associated with failed attempts to treat the free-surface term.
(In a future commit, I might remove the free-surface term entirely.)

- Reorganized the inner solution loop to include basal melting in addition to velocity
and continuity.

- Modified the entrainment so that it depends on the cell-center average slope and plume speed.

- Reverted to a first-order upstream method for computing the plume thickness at edges
in the continuity equation.