Added grounding-line fields and diagnostics
This commit included code written by Gunter Leguy to diagnose fluxes at grounding lines.
A grounding-line edge is defined as an edge with grounded ice on one side, and
either a floating cell or ice-free ocean on the other.

Four 2D fields were added to glide_vars.def:
- gl_flux_east  = flux through GL at east cell edges (kg m-1 s-2)
- gl_flux_north = flux through GL at north cell edges (kg m-1 s-2)
- gl_flux = sum of magnitudes of outbound GL fluxes for a given cell
- gl_flux_tavg = time average of gl_flux
The first two can be either positive or negative, depending on direction,
whereas gl_flux is positive by definition.

Also, a scalar field called tot_gl_flux was added; this is the total GL flux
over the CISM domain.

The 2D fluxes are computed in subroutine glissade_grounding_line_flux in glissade.F90.
The scalar total flux is computed in glide_diagnostics.F90.

Note that the fluxes are diagnosed from the velocity along with an approximate GL thickness.
In general, they will not agree exactly with the fluxes computed in the transport scheme.

This commit is diagnostic only; no answer changes