Changed value of the unphys_value parameter
I changed the parameter unphys_value from -999.d0 to -99999.d0.

Several fields are set to this value to identify them as physically unrealistic.
However, a bed topography of -999 m may be a real value.
Setting unphys_val to a more negative value reduces the chance of misdiagnosing
an actual value as unphysical.

Answers are unchanged.