Added an adaptive time step option based on the CFL number
The CISM time step is limited by a CFL requirement: Ice trajectories cannot
exceed one grid cell length in a single step. This requirement can be restrictive
in simulations that occasionally have large maximum velocities but usually have
modest maximum velocities (< 5 km/yr). Large velocities are often temporary,
resulting from long peninsulas of floating ice that are in the process of calving,
usually in the first few time steps of a simulation.  In these cases we would like
to allow the ice to calve without being limited to a small time step for the full run.

Previously, CISM has dealt with these situations by subcycling the transport;
that is, by calling the transport solver two or more times between calls to the
more expensive velocity solver. This subcycling gets around the CFL limitation,
but may be unnecesssary for all but a few time steps.

This commit adds support for adaptive subcycling. There is already a test during
each time step for advective CFL violations. Now the user can specify a CFL limit
(called adaptive_cfl_threshold), beyond which the transport is subcycled as often
as needed to satisfy the limit.

The check is done each time step. Typically, CFL violations (if they occur)
are associated with narrow floating peninsulas. During the advective subcycling,
the fast cells usually detach from the main shelf and are removed as icebergs.

A value of advective_cfl_threshold = 0.5 is usually sufficient to prevent
advective CFL errors from crashing the code. For the default value of 0.0,
there is no adaptive subcycling.

Also, I commented out the familiar diffusive CFL warning. For higher-order velocity solvers,
the diagnosed diffusive CFL limit has been found to be needlessly restrictive.

This commit is answer-changing if adaptive_cfl_threshold > 0.
For the default value of 0, answers are unchanged.