Bug fix for mask call in glissade_therm
In glissade_therm.F90, there was a call to glissade_get_masks in which the
optional floating_mask and ocean_mask arguments were included without a keyword
(i.e., 'ocean_mask = ocean_mask'). Since there are several optional arguments,
it is advisable to use a keyword with each argument.

This bug fix can be answer-changing for simulations with marine ice
and prognostic temperatures. Standard LIVV tests (including dome) are BFB.