Various modifications to support outflow boundary conditions
With an outflow boundary condition, any ice reaching the global boundaries
should gracefully exit the domain. This is different from periodic BCs (the default),
in which the ice leaving the southern edge of the domain enters the domain
from the north, and similarly for other directions.

Until now, outflow BCs have nominally been supported, simply by setting scalar
arrays to zero in cells outside the global boundary. However, velocities were
not treated correctly. For a correct treatment, the velocities along the
north, south, east and west edges of the domain should be computed on the appropriate
processor, but should not be passed to the opposite side of the domain.

This is tricky for staggered fields (e.g., velocity) along the southern and western rows
of the domain, which have a local index of i = staggered_lhalo or j = staggered_lhalo.
There were many loops in the code, supposedly over locally owned verticess, with
a lower limit of staggered_lhalo + 1. For outflow BC, these loops erroneously
omitted locally owned vertices with i or j = staggered_lhalo.

This commit includes the following changes:
- In the parallel modules (parallel_mpi and parallel_slap), I introduced new
  integer variables on each processor called staggered_ilo, staggered_ihi,
  staggered_jlo and staggered_jhi. These integers denotes the lower and upper
  indices for physically real, locally owned vertices.
- For processors lying along the south or west edge of the domain, staggered_ilo
  and/or staggered_jlo is set to staggered_lhalo (if CISM is run with outflow BCs).
  For other processors (and for all processors, if using periodic BCs), we have
  staggered_ilo = staggered_jlo = staggered_lhalo+1.
- I modified many loops in glissade_velo_higher.F90 and glissade_velo_higher_pcg.F90
  to use these new limits.
- I added or modified several subroutines in parallel_mpi and parallel_slap
  to handle outflow BCs correctly, using the new staggered_ilo/ihi/jlo/jhi values.
  Values along the north, south, east and west edges of the global domain
  are passed to neighbor processors as needed. Values at vertices outside
  the global domain are set to zero.
- I removed the logical variable 'fill_global_halos'. The procedure for filling
  halo cells is now based on whether the BCs are periodic or outflow.

I wrote some new code in subroutine glissade_test_halo (glissade_tests.F90)
to make sure the new staggered_parallel_halo subroutines act as intended.

A related change in the parallel modules: I added new subroutines for extrapolating
scalar fields to halo cells. With outflow BC, it can be useful to extrapolate
topography (topg) rather than set topg = 0.0 in halo cells.  This is because topg = 0.0
implies a land cell, whereas the edge of the domain for whole ice-sheet simulations
often consists of ice-free ocean.

I removed the global_sum subroutine, which was redundant with function parallel_global_sum.

In a couple of error diagnostics, I added calls to subroutine parallel_globalindex
to determine the global i and j indices of the problem cell.

I ran a dome problem (with modified basal BC to increase velocities) for 100 years with
both periodic and outflow BC. As expected, output is BFB for several decades, until
ice reaches the edge of the domain. From that point, the periodic run conserves mass
while the outflow run slowly loses mass, and the two runs start to diverge.

Also, I ran the same dome problem with outflow BC (1) with the serial build and (2)
on 1 core with the parallel build. Results from (1) and (2) are BFB, as expected.

Thus the outflow BC appear to be working as designed.

Results (including standard LIVV tests) are BFB with periodic BC.
Answers are generally different for outflow BC.