Added subroutine for cleaning up variables in ice-free cells
I added a subroutine called glissade_cleanup_icefree_cells.
This subroutine sets tracers to zero (or whatever is the appropriate value
for an ice-free cell) in all cells where thck = 0.

There are several places in the calving module where ice thickness can be
be set to zero. Rather than zero out tracers each time this is done,
I added a single call to the new subroutine in glissade_tstep, just after
the calving solve.

This commit might be answer-changing for some runs (if tracers were not
correctly zeroed out before this commit), but standard LIVV tests are BFB.

Note: As new tracers are added to the code, they should be added to
this subroutine. At some point it might make sense to put all such tracer
utility subroutines in one module.