Added a no-advance calving mask option
One of the calving options is to calve ice in all cells where calving_mask = 1.
Previously, calving_mask could either be read in from the input file at initialization,
or computed based on config parameters calving_front_x and/or calving_front_y.

With this commit, the mask can also be computed by setting calving_mask = 1
in all cells that are initially ice-free ocean. This option is used as the default
when calving_mask is not read in and calving_front_x/y = 0.

This option is similar to the no-advance mask, but it applies to ocean regions only.
It can be used to prevent the calving front from advancing beyond observed locations.

Answers are unchanged unless this calving option is set.