Modified iceberg removal, added cull_calving_front option
This commit includes some minor logic changes in the iceberg-removal algorithm.

The algorithm can now be described as follows:
(1) Mark all cells with ice (either active or inactive) with the initial color.
    Mark other cells with the boundary color.
(2) Seed the fill by giving all active grounded cells the fill color.
(3) Recursively fill all cells that are connected to filled cells by a path
    that passes through active cells only.
(4) Repeat the recursion as necessary to spread the fill to adjacent processors.
(5) Once the fill is done, any cells that still have the initial color and
    are not on land are considered to be icebergs. They are removed.

Notes:
(1) The recursive fill applies to edge neighbors, not corner neighbors.
    The path back to grounded ice must go through edges, not corners.
(2) Inactive cells can be filled (if adjacent to active cells), but
    do not further spread the fill.
(3) Grounded cells that still have the initial color are not removed.
    They are considered harmless. (It would probably be OK to remove them,
    but this would be answer-changing for the dome problem.)

This process not only identifies floating cells that are disconnected
from the main ice sheet (because ocean cells lie in between), but also
removes inactive floating cells that are not adjacent to active cells.
Thus the layer of inactive calving-front cells can be only one cell thick.

Also, I added a cull_calving_front option, which is false by default.
If set to true, it removes all calving front (CF) cells at initialization.
The process can be repeated, removing multiple layers of CF cells, by
setting ncull_calving_front > 1.

This is a crude way of removing skinny peninsulas, which tend to be unstable.
For instance, suppose we have a peninsula that is two cells thick.
Each cell on the peninsula is a CF cell, so the entire peninsula will be removed.

I am not sure we will want to keep this option. It may be sufficient to
use iceberg removal (which will remove peninsulas two cells thick when the
subgrid CF scheme is used) combined with an adaptive CFL (which usually can
handle large velocities for a few time steps, until fast cells drift away
and are removed). But it is there for now as a last resort.

Answers are unchanged for standard LIVV tests.
Answers will change for runs with dynamic marine ice sheets, including MISMIP+.