Reworked eigencalving and thickness-based calving
After much trial and error, I settled (for now) on a new eigencalving algorithm.
The procedure is:
(1) In dynamically active cells, compute the determinant of the 2D strain rate tensor,
which is equal to the product of its two eigenvalues ('strain_rate_eigenprod').
Save the determinant where it is positive and both eigenvalues are positive.
(2) Extrapolate positive values of the determinant from active floating cells to
adjacent inactive calving-front cells.
(3) Use the Levermann et al. (2011) scheme to compute a calving rate in each cell
with positive spreading rate.
(4) Diagnose the effective thickness in calving-front cells.
(5) Based on the effective thickness, convert the calving rate to a thinning rate
in CF cells. Calve ice in these cells.
(6) If the thinning rate is large enough to remove more ice than is present in a
calving-front cell, then remove all the ice in the cell and apply a small amount
of thinning to the upstream interior cell. The thinned upstream cell will usually be an
inactive CF cell during the following time step. In this way the CF smoothly retreats.
(7) Apply thickness-based calving as described below. In other words, some initial
thinning is applied using the eigencalving algorithm. When the ice is thin enough,
thickness-based calving applies the coup de glace.

Note: The eigencalving is a bit counterintuitive in that thinning generally
occurs in inactive cells where we cannot compute the spreading rate directly.
But numerically, it helps to confine the thinning to inactive cells, because
otherwise the eigencalving would lead to large surface gradients and driving stresses.

I also reworked the thickness-based calving. Thinning is now applied to CF cells
whose effective thickness (H_e = thck_calving_front) is less than calving_minthck.
The thinning rate is given by
    dH_e/dt = -(Hc_min - H_e) / tau_c  where Hc_min > H_e
    dH_e/dt = 0 elsewhere
Here, tau_c = calving_timescale controls the rate of thinning/calving.
Note: Typically, thck_calving_front > thck. Ice does not thin/calve simply because thck is small.

If the thinning rate is large enough to remove more ice than is present in a cell,
some additional thinning is done upstream (as described above for eigencalving).

After eigencalving and/or thickness-based calving, there is a cleanup step.
In this step we look for CF cells with thck > thck_calving_front.
(That is, ice at the CF is thicker than upstream ice, which is not realistic.)
In these cells we set thck = thck_calving_front and calve the difference.
During the next time step, these cells will usually be active CF cells
(since CF cells are active when thck >= thck_calving_front).

Both thickness-based calving and eigencalving rely on the subgrid calving scheme for stability.
If either of these calving options is turned on and which_ho_calving_front = 0,
the code automatically sets which_ho_calving_front = 1, and reports the change in the log file.

I removed the old thickness-based calving scheme which relied on a protected ring
of CF cells around active ice. The new scheme is similar in concept, but (as pointed
out by Albrecht et al.) better captures shelf dynamics by using thck_calving_front
to give smooth, gradual changes in CF thickness.

I'm not entirely satisfied with either eigencalving or thickness-based calving.
In particular, it is hard to tune the eigencalving so that real calving fronts
(e.g., in Antarctica) stay in the observed location. Big ice shelves (e.g, the Ronne-Filchner)
tend to advance too far seaward, and small shelves (e.g., Thwaites) tend to retreat.
At some point I may try a similar scheme based on stresses rather than strain rates.

This commit is answer-changing for eigencalving and thickness-based calving,
but otherwise is BFB.